apply the refactoring for the group features

+++ server/src/features/group/admin/service.js
@@ -1,70 +1,83 @@
 import Group from '../../../models/group.model.js'
 import User from '../../../models/user.model.js'
+import ERRORS from '../../../../config/_errors.js'
 
-export const createNewGroup = async (userId, data) => {
+export const createNewGroup = async (user, data) => {
    const { name, username, image, info } = data
+   if (!name) return ERRORS.INVALID_CREDENTIAL
+
    const group = new Group({
       name,
       username,
       image,
       info,
-      owner: userId,
-      members: [userId],
+      owner: user.id,
+      members: [user.id],
    })
    const newGroup = await group.save()
 
-   if (newGroup)
-      await User.findByIdAndUpdate(userId, {
-         $addToSet: { groups: newGroup.id },
-      })
+   if (!newGroup) return ERRORS.SERVER_FAILED
 
-   return newGroup
+   user.groups.addToSet(newGroup.id)
+   await user.save()
+
+   return { data: newGroup }
 }
 
-export const findGroupAndUpdate = async (userId, groupId, data) => {
-   const { name, username, image, info } = data
+export const findGroupAndUpdate = async (user, data) => {
+   const { groupId, name, username, image, info } = data
    const update = { name, username, image, info }
 
    const group = await Group.findById(groupId)
-   if (!group) return null
+   if (!group) return ERRORS.NOT_FOUND
 
-   if (!(group.owner.equals(userId) || group.admins.includes(userId)))
-      return null
+   if (!(group.owner.equals(user.id) || group.admins.includes(user.id)))
+      return ERRORS.FORBIDDEN
 
    Object.assign(group, update)
    const updatedGroup = await group.save()
+
+   if (!updatedGroup) return ERRORS.SERVER_FAILED
+
    updatedGroup.members = undefined
-   return updatedGroup
+   return { data: updatedGroup }
 }
 
-export const findGroupAndDelete = async (userId, groupId) => {
-   const group = await Group.findOneAndDelete({
-      _id: groupId,
-      owner: userId,
-   })
-   return group
+export const findGroupAndDelete = async (user, data) => {
+   const { groupId } = data
+   if (!groupId) return ERRORS.INVALID_CREDENTIAL
+
+   const group = await Group.findOneAndDelete({ _id: groupId, owner: user.id })
+   return { data: group }
 }
 
-export const addAdminToGroup = async (userId, groupId, adminId) => {
+export const addAdminToGroup = async (user, data) => {
+   const { groupId, adminId } = data
+   if (!groupId || !adminId) return ERRORS.INVALID_CREDENTIAL
+
    const group = await Group.findById(groupId)
-   if (!group) return null
-   if (!group.owner.equals(userId)) return null
+
+   if (!group) return ERRORS.NOT_FOUND
+   if (!group.owner.equals(user.id)) return ERRORS.FORBIDDEN
 
    group.admins.addToSet(adminId)
    const updatedGroup = await group.save()
 
    updatedGroup.members = undefined
-   return updatedGroup
+   return { data: updatedGroup }
 }
 
-export const removeAdminFromGroup = async (userId, groupId, adminId) => {
-   const group = await Group.findById(groupId)
-   if (!group) return null
-   if (!group.owner.equals(userId)) return null
+export const removeAdminFromGroup = async (user, data) => {
+   const { groupId, adminId } = data
+   if (!groupId || !adminId) return ERRORS.INVALID_CREDENTIAL
 
-   group.admins.pull(userId)
+   const group = await Group.findById(groupId)
+   if (!group) return ERRORS.NOT_FOUND
+   if (!group.owner.equals(user.id)) return ERRORS.FORBIDDEN
+
+   group.admins.pull(adminId)
    const updatedGroup = await group.save()
 
    updatedGroup.members = undefined
-   return updatedGroup
+   return { data: updatedGroup }
 }
